split extension constants to new file
