+++ src/client/src/containers/AppServiceModal/index.tsx
@@ -10,7 +10,7 @@ import AppNameEditor from "./AppNameEditor";
 import SubscriptionSelection from "../../components/SubscriptionSelection";
 import { InjectedIntlProps, injectIntl } from "react-intl";
 import buttonStyles from "../../css/buttonStyles.module.css";
-import { WIZARD_CONTENT_INTERNAL_NAMES, KEY_EVENTS, AZURE, EXTENSION_COMMANDS, SERVICE_KEYS, AzureResourceType } from "../../utils/constants/constants";
+import { WIZARD_CONTENT_INTERNAL_NAMES, KEY_EVENTS, AZURE, SERVICE_KEYS, AzureResourceType } from "../../utils/constants/constants";
 import styles from "./styles.module.css";
 import { AppState } from "../../store/combineReducers";
 import { IAppService } from "../../store/userSelection/services/appService/model";
@@ -24,6 +24,7 @@ import LocationSelection from "../../components/LocationSelection";
 import { ReactComponent as ArrowDown } from "../../assets/chevron.svg";
 import { AppContext } from "../../AppContext";
 import ResourceGroupSelection from "../../components/ResourceGroupSelection";
+import { EXTENSION_COMMANDS } from "../../utils/constants/extension";
 
 interface IStateProps {
   isModalOpen: boolean;
