+++ src/client/src/containers/RightSidebar/ServicesList/AppServiceSelection/index.tsx
@@ -4,13 +4,14 @@ import SidebarItem from "../SidebarItem";
 import { IAppService } from "../../../../store/userSelection/services/appService/model";
 import { ReactComponent as EditIcon } from "../../../../assets/edit.svg";
 import styles from "./styles.module.css";
-import { KEY_EVENTS, EXTENSION_COMMANDS } from "../../../../utils/constants/constants";
+import { KEY_EVENTS } from "../../../../utils/constants/constants";
 import { injectIntl, InjectedIntlProps } from "react-intl";
 import { sendTelemetry } from "../../../../utils/extensionService/extensionService";
 import { removeAppServiceAction } from "../../../../store/userSelection/services/appService/action";
 import { AppContext } from "../../../../AppContext";
 import { openAppServiceModalAction } from "../../../../store/navigation/modals/action";
 import messages from "./messages";
+import { EXTENSION_COMMANDS } from "../../../../utils/constants/extension";
 
 interface IProps {
   appServiceSelection: IAppService | null;
