+++ src/client/src/mockData/extensionModules/mockAzureModule.ts
@@ -1,4 +1,4 @@
-import { EXTENSION_COMMANDS, EXTENSION_MODULES } from "../../utils/constants/constants";
+import { EXTENSION_COMMANDS, EXTENSION_MODULES } from "../../utils/constants/extension";
 
 import * as mockData from "./mockAzureModuleData";
