+++ src/client/src/mockData/mockVsCodeApi.ts
@@ -1,5 +1,4 @@
 import {
-  EXTENSION_COMMANDS,
   DEVELOPMENT,
   TEST
 } from "../utils/constants/constants";
@@ -7,6 +6,7 @@ import {
 import * as mockAzureModule from "./extensionModules/mockAzureModule";
 import * as mockLoggerModule from "./extensionModules/mockLoggerModule";
 import * as mockGenerationModule from "./extensionModules/mockGenerationModule";
+import { EXTENSION_COMMANDS } from "../utils/constants/extension";
 
 /**
  * Models the functionality of acquireVsCodeApi() from vscode for use
