+++ src/client/src/utils/extensionService/extensionService.ts
@@ -1,8 +1,9 @@
 import { IVSCodeObject } from "../../types/vscode";
 import {
-  EXTENSION_COMMANDS, EXTENSION_MODULES, WIZARD_CONTENT_INTERNAL_NAMES, PAYLOAD_MESSAGES_TEXT
+  WIZARD_CONTENT_INTERNAL_NAMES, PAYLOAD_MESSAGES_TEXT
 } from "../constants/constants";
 import { ILoggingPayload } from "../../types/logger";
+import { EXTENSION_COMMANDS, EXTENSION_MODULES } from "../constants/extension";
 
 const postMessageAsync = (command: string, paramsMessage: any, vscode: IVSCodeObject, scopeId: number = Math.random())=>{
   const promise = new Promise<any>((resolve) => {
